test: refactor retry helper and make it generic
Initially, we declare the requests retry helper within the authentication
tests but this helper could be made generic such that other tests could
utilize it.

Now, the request retry helper is defined in `test/utils.py` before being
imported into the authentication tests.
